Implement user friendly interface for load-cell driver
